+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 * Change data viewing 'progress' bar to have a minimum width of 3px to make it more visible.
 
+* Issue **#2462** : Made changes to stop DB connections being used within the context of other open connections.
+
 * Uplift LMDBJava to 0.8.2 to fix LMDBJava cursor comparator bug.
 
 * Issue **#2464** : Add `leakDetectionThreshold` to the hikari pool config.